Implemented FPS display
The FPS are now displayed on the screen using the InfoUI.

The InfoUI prefab contains the entire setup for the UI to display information. As of now, it only displays the framerate of the program.

The InfoUIController class controls the logic to update the framerate every so often.

The fps display text is not updated every single frame to prevent wasting resources.