Added light to init scene and skybox material
In Unity, the main / init scene (whatever scene is loaded first) NEEDS to have the same lighting and skybox setup as the default scene. Otherwise, the lighting system breaks, and using SceneManager.Load() will never load lighting correctly.

By adding a default light to the init scene and restoring the skybox, this issue has been fixed.

This issue has been known for years, but unity has yet to patch it... it is what it is...