Temporarily disabled SetActiveScene call
The SceneLoadingManager's method LoadScene() no longer calls SceneManager.SetActiveScene().

This is because it actually is not required, altough it is a nice formality. The problem comes from the fact that Unity requires at least 1 frame to pass between the LoadScene() call and the SetActiveScene() call for Unity's base SceneManager to be able to register that the selected scene has been loaded.

The call to SetActiveScene() requires that the active scene we're setting is actually registered as loaded.

This call is not necessary as of now since we're not loading multiple scenes and enabling and disabling some of them as we go. Right now, we only load 1 single scene at a time, so this call is unnecessary.