Improved camera manager
The camera manager is now a Singleton<> and implements the IManager interface.

The camera rotation is now updated as well to match the target transform.

The camera manager now makes sure that the active camera and target are not null before updating the camera position and rotation.

TODO : The loading manager has to call ResetManager() on camera manager to make sure that we don't have dangling references to objects that no longer exist on the scene when loading a new scene.