Implemented basic event driven input system
The input manager now makes use of Action<> events as callbacks for the input system.

Dereferencing function pointers is slightly slower than calling the function in raw as the old implementation did on my previous repo, but this implementation is far more flexible and allows for a greater level of separation of concerns and logic encapsultaion, which makes extending the input system and implementing UI related inputs trivial.

TODO : Make some of the input functions public for use from the player UI later on.