Fixed bug with camera being null in player controller
The player controller function SetLookToPoint() would throw an error if the active camera referenced by the camera manager was set to null at some point.

This has now been fixed by adding null checks, allowing malformed levels with no camera to work without breaking the game.

This has also fixed another issue where, if the player camera's script was executed a frame after the player's script, then the function would generate an error once and work for the rest of the runtime. This was a minor inconvenience, but it is better to be safe than sorry.