Added display text area to console UI
This display text area will be used to print the cmds to the console.
Under the hood, this is handled by CmdPrint(), which also calls the debug manager's Log function.

NOTE : The text is set to be adjusted from the bottom so that we can get the most easy and automatic setup for infinite text printing on the console, but this means that it will never get flushed unless we manually call the clear command. This can lead to some performance issues if someone goes overboard printing shit to the console.