User cmd input is now printed before running cmd
The commands written by the user are now printed before running or even parsing the command.

This is done to allow the user to see exactly what is it that they typed into the console prior to anything being executed, that way they don't lose commands and they can spot typos or whatever.