Fixed case sensitive cmd str issues and added delete cmd
Fixed issues where specific case sensitive strings would be passed to lower case entirely on the command handling functions, making it impossible for the command to be executed correctly.

This has been fixed by making it so that only the base command is passed to lower case, and the arguments for the command are kept as they are, which makes args that are case senstitive work properly.

A delete command has been added to allow the user to delete game objects. This is a WIP for a kill command for entity testing.

TODO : Implement some basic entity system for a command inspited by JK's use command.