Added debug command and extended CmdError
There is now a debug command which allows enabling and disabling debug mode during runtime.

The CmdError command has been modified to have a set of predefined error messages according to the error type. This list is bound to be changed and even removed, restoring the old system. It's still WIP, so yeah.