Fixed bug in Remove() method from ElementQueue
The remove method used to look from the removed element to the last element minus one during the shifting phase.

This is wrong since the total element count is previously reduced prior to running the shift loop, which made it so that elemental combinations would give unexpected results.

This has now been fixed by looking through from i to Count.