Added method to get element count to ElementQueue
The element queue class has a publicly accessible public get property named ElementsCounts, which contains an array of integer values, each containing the number of instances of a given element within the element queue.

Accessing each element's count can be annoying to write out every single time due to having to explicitly cast enums to integer types in C#, so a method has been provided to make this process easier to deal with.

The method GetElementCount() returns the count of a given element.
If the elemnt does not exist, it simply returns 0.