Updated player controller logic for spell casting and movement
Separated walk and gravity movement in the UpdateMovement function so as to allow gravity to keep affecting the player while all other types of movement are disabled (during can't walk states such as casting spells)

Updated the cast spell function and separated it into right click down and right click up events. This will allow spells that are casted over time while the click is held down to be easier to implement.