Fixed issue where LineRenderer was used before it was set
The debug line renderer controller made use of Unity's GetComponent<> function to get the LineRenderer component at the start.

This is a problem because the order in which start is executed is not guaranteed to be as soon as the object is created, which means that the debug manager could be referencing null data.

This could be partially fixed using awake, but the best solution (and the one implemented) is to set the line renderer as a parameter on the inspector within the debug line renderer prefab.