Created player data manager
The player data manager class is responsible for holding information and references to data of the player so that it can be globally accessed at any point.

This type of information would usually be found within a game manager of sorts, but it has been made into its own class to separate player data getting from gameplay specific logic, as well as allowing some of the player data to be persistent between levels / sessions in the future.