Modified spell caster to add beam spell color
WIP : The color is calculated with the average of all elements within the element queue.

This calculation needs to be improved to obtain more visually appealing colors, as well as making it possible to configure other visual aspects of the beam and other spell types.

TODO : Change it so that the elements counts array is all that has to be passed to the spawned spell on creation.