Fixed issue where HUD would not update on level change
The player HUD used to maintain the old element queue if the player returned to the menu while some elements were queued up.

On load, the elements would still be there visually, but not within the real element queue, so this issue was just visual, due to the way that the queue is updated.

This issue has now been fixed by updating the queue when the load level function is called.