Deduplicated null checking in player UI
The player UI controller now uses the AllComponentsAreValid() method for UI updating, which means that none of the specific update methods make their own null checking.

This has the downside that we can't have partial UI updates, if one system breaks, everything freezes, but that makes it so that we at least do not require checking for null more than once for every single component that is required by more than one component within the UI.

Yes, it looks like a dumb micro optimization, but the point is not just the performance (an extra null check is basically free), the point is deduplicating code and reduing points of failure, as well as making maintanance easier.