Modified console to scroll to bottom when a cmd is ran
Running a command now makes the console automatically scroll to the bottom, so that the end of the output generated by a command's execution can be visualized on screen easily.

This is standard behaviour for most consoles, where the view jumps to the cursor location when a command is executed or the enter key is pressed.