Added cls command
The cls command is simply an alias / copy of the clear command. It only exists for preference and ease of use.

Windows users are more likely to naturally try using cls, while Linux users / users of Unix based systems are more likely to use clear. It's just nice to allow both to work, that's it.

It's not a big deal, but this feels like one of those QoL features for console commands users.