Created SpellCaster Prefab
This prefab will contain all of the components required for the SpellCasterController to be able to properly perform all of its logic and functionality.

This prefab can be placed and attached to any entity that will perform spellcasting for ease of control, and, in the future, for automatic spell pooling support out of the box.