Fixed PlayerDataManager component getting for spellcaster
The spell caster controller was wrongfully trying to be gotten as a component attached to the player's gameobject. This is not the case, as the spell caster is a component attached to a child object that is located within the player's prefab.

This has now been fixed by making use of the GetComponentInChildren<>() function instead.