Fixed typo in spell caster form shield case
The shield case was using early returns when it should be using continues to skip to spawning the next wall once the final wall type has been determined.

This caused the walls to only spawn one single wall and never empty the element queue.

This problem has now been fixed.