Implemented SpellBaseController
The spell base controller class is now implemented.

It implements the ISpell interface and defines the functionality to correctly choose the color of the spell.

It contains a virtual function that can be used by each specific spell implementation to define how to physically handle color updates and what parts of the spell's prefab will have their color changed each time a color update takes place.

TODO : Implement system to calculate damage based on the elements that are part of the spell.