Changed default spray particle color to white
This makes it easier to tint the spray particle effects from the script, as the hue no longer depends on the original color being reddish / orangeish due to the particle being originally a fire particle.

TODO : Add system to change the particle's sprite to snow, steam and electricity according to the elements used.