Added borders for button sprite slicing
With the newly added borders, the buttons in the UI can now scale to any length without deforming the details on the edges of the sprite, allowing us to have buttons of any size without having to make custom sprites for every possible size.

This fixes all sprite scaling issues for UI elements.