Created AccountManager
The account manager will serve as a global access point in the code to allow manipulating the user account data and performing account related actions such as registering, logging in, modifying and deleting the account.

The account field stored within the account manager holds the data once the user has been logged in so as to allow certain actions to be streamlined.

NOTE : The idea of holding the password in memory is dangerous and is bound to be changed, as most actions are planned to require the user to input the password manually every single time rather than automatically allowing the user to perform such actions without the site asking for credentials again. Altough, sending forum messages in the future, changing inventory items, etc... all of those actions would be annoying if they required the user to input the password every single time, so this may need to be modified to make it safer, but the password field in the Account class has potential for staying if the right conditions are met.