Added MakeRequest() example to ConnectionManager
Added a working example that demonstrates how to make use of the MakeRequest() method from the ConnectionManager instance.

This method handles all of the possible cases when performing an async request to a server.

Any code that relies on the reply should be ran through delegates, which would most likely point to lambdas, so as to guarantee that the code does not run without having received a response first.