Implemented AccessAccount() method in AccountManager
The method performs a GET request to access the account data.

Note that, even if the user manipulates the game on the clientisde to make it think that it is logged in, all of the other functionalities offered by the server should ALWAYS check for credentials, which would make it so that an user faking being logged in would be completely harmless, as it would be a purely visual thing that only the client who performed the manipulation in the first place would be capable of seeing on their screen.