Created LanguageManager
The LanguageManager class will be in charge of containing all of the language data for the different languages supported by the game.

This is a WIP and will be further refined later on in later stages of the project, but this time I'm setting it up relatively early on so that I don't miss any UI elements when adding multi-language/localization support.