Replaced LanguageManager with GameLanguage static class
The LanguageManager class has been replaced with a static class GameLanguage.

The original idea was to be able to customize the language loc string entries from Unity's instpector panel on the GameObject that has the LanguageManager component.

This is a very dangerous idea, as changing the structure of any of the classes involved or renaming the input variables could lead to all of the localization entries getting deleted, which could erase a ton of work in a pretty dumb way.

In the future, this could be reworked to be a manager again if I figure out how to implement reading JSON files as inputs for the language data so that it could be easier to modify the language data without changing game code, which would also make modding easier in the future, but that requires file access in a Web build and I am not sure what the correct way to do that in Unity is.