Reworked language system to use enum instead of strings
The language system now uses an enum rather than a string for languages.

This deacreases moddability in the future, but it's not like we want this game to be modded as of now, so this doesn't matter that much.

The usage of enums decreases language lookup times, altough it was already short enough.

TODO : Change some of the dicts to be plain arrays since we're using enums with integer values and lookup times would be O(1) anyways.