Modified language system to wrap around when using integers
The SetLanguage(int) method now wraps around the language when given an integer value that is out of range.

This might not be a wise decision in many contexts, but in this case it is good enough because it allows the language config menus to have an easier time wrapping around the selected language during configuration.

It also allows the language system to never be able of entering an invalid language state when using an integer language index, but it also does not allow knowing if the index is valid or not unless you perform the check manually and do error handling on your own.