Modified language system
The language system now has new language keys, as well as having a new system for integer based language setting.

The wrap around code did not work correctly because the modulo operator does not work the same way in Unity's version of C# as it does in base .NET or C or C++, and I can't be assed to figure out why Unity does things different for the fucking modulo operator, so yeah.