Modified InputManager to detect inputs over UI buttons
The input manager no longer sends input events when the input action is pressing an UI button object. This prevents the player from moving or casting spells accidentally when pressing the buttons on the UI.

This is specially handy for mobile users, altough the pointer impl has only been tested on PC.

TODO : Figure out if this works on mobile platforms.