Modified PlayerController to notify death and display death screen
This is full of hacks and is in severe need of a refactor.
Most of this logic should be handled by a game manager or some other kind of manager in the future. Maybe the UI manager itself.
Also, this would be a hell of a lot better if we had a PauseManager, then the logic could be entirely independent and work through getting events from PlayerDataManager's fetched player health controller.