Created VersionManager singleton persistent class
The VersionManager class' purpose is to guarantee that the version of the game is easily confiugrable from Unity's inspector panel, as well as allowing any part of the code to get access to the current product version for whatever purposes.

Note that the GameVersion struct can also be used to perform save data translation functions in case that the save format structure changes between versions by adding incremental save methods.