Added Shader Graph to project
Added the shader graph to the project to facilitate shader editing.
Note that while the description of the shader graph package says it only supports URP and HDRP, it in fact also supports the standard render pipeline, which is what this project uses to obtain better performance on mobile devices, as well as allow supporting a wider range of older devices during runtime.