Modified beam spell prefab to improve visual quality
The beam spell prefab has been modified to improve the visual quality and eye candy.

It now makes use of the new beam material and shader, adding a moving beam effect as well as an inner glow.

The size of the beam's line renderer, it's UV scales and the scale of the particle effects have all been updated to increase visual appeal.