Implemented backbone for address fetching system
The address fetching system has been partially implemented.

The connection manager can now use a GET request to obtain the server addresses table from github and obtain a JSON file with the data inside.

If the operation is successful, the server will be configured to use the specified address.

If the operation fails, the server will use the custom address ip and port as a fallback / default and work with it instead, probably not being able to connect to the server in any way but allowing the game's systems to properly function even if the server is down or if the user has no internet connection.