Added character controller based movement to NPCs
The NPCs now make use of the character controller to define their movement. This way, the AI pathing and the physical movement can be separated, allowing NPCs to be thrown off course through knockback forces such as explosions or other spells without breaking the AI and without allowing NPCs to clip through walls when pushed around.