Added nav mesh distance check for AI position warping
The AI controller's logical position is now only warped if it is close enough to a valid nav mehs.

Do note that the logical position will continue to walk on top of the last valid nav mesh surface toward the target position and then be warped back once the entity finds a new valid location. This should be fixed later but it is ok for now.