Implemented AI enabled flag for test NPC
The hasAI / AIEnabled flag now works and is fully implemented, allowing the AI and pathing systems of the NPC to be fully disabled.

For now, AI logic is isolated within each specific function. This has a small overhead due to the extra checks performed by each function every frame. Also it localizes the checking to each individual function. In the future, this could be refactored to have the if statement in the update method and then call each AI and non-AI functions within the corresponding blocks.