Implemented AISensorArea controller
This controller is in charge of sensing when the player enters an area around the sensing entity.

The sensor will alert any listeners through the OnSense event if the player is within the sense area. The only condition is that the player's distance is within the radius / bounds defined by the area.