Created AISensorSight controller
A sensor controller that inherits from the base sensor class. The purpose of this sensor class is to detect / sense when the player is within the sight of the NPC.

Works similar to the way the area sensor works, only that it also performs a visibility check through physics raycasting

TODO : Fix situation where the player enters the sight area while hidden and then walks in front of the NPC but cannot be detected since the OnTrigger method is not called again. This can be fixed on the base sensor class by sensing on trigger stay.