Added owner system to health modifier controller
The owner system allows the health modifier controller to ignore a game object if they are marked as the "owner". The owner is to be used to select the gameobject that contains the health component which will be used by the NPC, Player or Entity that is understood at the gameplay level to be the "owner" of the attack / damage area.