Fixed issue with Repeat mode damage application
For some reason, exposing float variables on Unity's inspector breaks float comparison with >= and <= as well as > and <, all comparisons seem to compile to == for some reason. This is clearly a problem on the compiler level and the temporary fix / patch is to add a temp var to store a copy of the exposed value to before performing the comparison.