Modified AI Controller logical position system and added stopping distance
The AI controller's logical position system has now been simplified to depend on a behaviour in Unity where adding a NavMeshAgent compoent along with a CharacterController component gives the expected behaviour out of the box, so we do not need to do by hand all of the work the old UpdatePathing() method used to do, as the logical and physical positions are now tied together.

This behaviour does not appear to be documented officially but it is consistent across versions, so it is going to be the solution used for now.