Added new long button sprite textures
Long buttons that are vertically aligned now have a new texture with higher detail and better eyecandy.

NOTE : Some of the smaller edge details could have some problems in higher resolutions, maybe increase the thickness of the decor edge to make sure the detail is not too fine / thin? That way we could prevent aliasing.