Created ISpellCaster interface for spell caster controllers
The SpellCasterController script is going to be modified to be a coordinator of multiple specific SpellCasterTControllers whose purpose will be to control one specific type of spell.

This interface serves as a contract that enforces implementing all of the required functionality for a SpellCaster controller of any given spell type.

It is most likely that each spell type / form will have a corresponding specific spell caster implementation.