Disabled Form handling from ISpellCaster
The spell caster interface should not have any requirement for form handling, since each specific spell caster implementation should be the one that already handles a specific form + element combination.

The spell caster controller system (not to be confused with spell caster base controller) should be the script in charge of determining the form of the spell.

NOTE : The caster should maybe be renamed to SpellSystemController or something to that effect to prevent confusion in the future.