Changed stop casting handling call order
The new call order has been chosen so that the stop casting spell methods can make use of the time accumulator variable to obtain how long the spell has been casted for by the time the spell casting was stopped.

This is specially useful for projectile spells, which require this cast time accumulator for projectile speed / force calculations.