Changed player's wall spawning height
The increase in the player's mesh height location to match the character controller's hitbox height has caused the spell caster's height position to change as well.

This increase in height has caused the walls to start spawning floating in the air.

This has been fixed by lowering the height at which the wall spell spawn transforms are located.