Updated SpellCasterController to use configurable element queue size
The element queue size can now be configured on the inspector panel to set a custom amount of max elements on the element queue for a given spell caster.

Since spells calculate their effects based on the total number of elements of the input element queue, we don't have to change any code on the spell controller classes, since those work with dynamic queue sizes out of the box.