Added basic element particle visibility handling
The spell caster controller now has a function to update the visibility of the element particles in the queued up spell.

Each element will display one floating particle around the spell caster's origin.

NOTE : This code updates every frame rather than on certain events such as adding and removing elements, so it may not be as performant as it should be. Check this out later and find a better way to solve this problem if possible.