Created SpellEmitterController script
The spell emitter controller script is in charge of controlling the logic of emitter spells.

Emitter spells are just visual representations of elements, and they do not have any other form of logic related to spell casting.

They simply inherit from SpellBaseController so that they can reuse the color setting logic without moving it to the magic manager.