Moved initialization code from Start() to Awake() on DebugManager
This change will allow systems that depend on the debug manager to run code on Start(), as the debug manager will be guaranteed to be initialized.

This way, we can skip adding null checks every single frame, which would obtain the same result, by allowing the game to enter a valid but ugly state in which calling SetDebugEnabled() would change debug enabled to true, but the line renderer disabling code would not be called since the line renderers list would still be empty. This state would be fixed on the next frame, but the null checking would still have to run every frame before calling DisableLineRenderers(), which on the long run could have a large performance impact.

It has been determined that it is better to actually handle initialization order correctly rather than looking for hacky workarounds. It is better to accept the fact that some Managers will need to be initialized on Awake() and others on Start() than letting them sort themselves out over 2 or 3 frames by adding a permanent null checking overhead to every function when we should instead guarantee that the systems will never have null values.

NOTE : Maybe an InitializationManager class would be good to have, so that we could manually call Init methods in the order that it is required for the systems to work properly?