Moved initialization code from DummyManager to InitSceneController
The screen orientation adjustment code has been moved from the DummyManager to the more logical init code within the InitSceneController.

The Init scene controller now has a macro that checks if the code is being compiled in editor or on a final build, making sure to always disable debug mode by default when starting the game. This way, we can prevent issues where we forget to disable the debug mode on release builds.