Moved dynamically tileable materials to a subdirectory
NOTE : There is a difference between tiling and dynamically tiling materials.

Tiling materials tile with a specific amount of repetitions no matter the scale of the object.

Dymically tiling materials allow the material to keep a consistent tiling ratio when scaling the object, which makes it great for reuse on many different objects with custom UVs and different scales, maintaining a consistent visual texture scale across all objects