Updated TriggerController to use a trigger count value
The trigger count value allows the user to specify how many times the trigger can be activated.

If the value N is positive, then the trigger will activate N times.

If the value N is 0, then the trigger will no longer be triggerable.

If the value N is negative, then the trigger will be capable of being triggered infinite times.