Modified TriggerController to use a single event and boolean config
The old system would force trigger users to specify multiple times the same event on different UnityEvent handlers on the inspector panel if a trigger was capable of triggering events on enter, on exit and on stay.

This made it complicated to configure scenarios where a trigger would need to have different behaviour and trigger counts related to each event. For example, a trigger that has a count N on enter, a different coutnt M on exit and an infinite count on stay. This would have been impossible to configure with the old system.

With the new system, a single gameobject can have more than one trigger component, each with a different event configured to run on multiple events or a single event or none at all (these types of triggers could be used exclussively through code by being called with the Trigger() method and adding events with OnTrigger += code;), each with their own name, count and specific config and events.

This makes the new system more flexible and capable of building more complex systems and structures within the level editor.