Created controller script for PopUp menus
This controller script is in charge of controlling the basics of the logic of pop ups.

PopUps will be small windows that appear on screen on top of the UI in menus (and hopefully NEVER in game), to display some kind of critical or important information, such as notifying connection errors when attempting account registration or other such network operations.

PopUps can be used for other purposes, but this is the main reason they will exist for as of now.