Created GameManager script
The GameManager script will be in charge of managing game logic and game mode specific functionality.

As of now, it is a non-persistent Singleton, which means that each level should have its own game manager. The game should be capable of running without the game manager being present, otherwise the design would become too coupled and dependant on this system existing even when it is not required.