Created SettingsManager script
The settings manager will be in charge of holding a global access point for a container with the chosen configuration selected by the user.

The settings manager internally holds the functionality to change the settings of the game, which means that the settings can me modified through commands or other means by accessing this manager's interface, rather than hardcoding in place everywhere that it is required the same functionality multiple times. For example, the current language selection implementation is coded in place on the Cmd controller and the settings menu controller. This logic can now be properly encapsulated into a class that manages it and serves as a global access point for performing said modifications.