Implemented basic system for score fetching
The score is not automatically updated on the local UserAccount instance.
This means that if the score is updated, when the player goes to their profile to check it, they will still see the old value.

A solution would be to implement fetching when a game is finished, but that would not account for multiple people using the same account, or the account running at the same time in multiple different computers, so that would lead to wrong results.

The chosen solution is a simple system where the data is fetched every time the profile is loaded, guaranteeing that the information displayed is always up to date.