Made GameManager a non persistent singleton
It has been decided to go back to the previous idea where the game manager exists on each level rather than it being a global system that persists across scenes.

This allows triggers in the scene to directly access game manager functionality. It also allows the game manager's logic to be reset.

Again, as a side note, it should be clear that the game should be coded in a way such that it should be capable of running without a game manager instance being present in the scene.