Added player score controller to PlayerDataManager
After finishing part of the game's implementation, it has finally become possible to come back to using the money and score controllers for points management.

This means that now, all of the game's systems will properly function even if no game manager is present in the scene (eg: with the game manager system, the player UI will no longer update if it is not on the scene. This is no longer the case.)