Created Fixed localized text controller
The fixed localized text controller script is a component that is in charge of localizing a specific text using the specified language as input. This is mostly useful for displaying the language config, since the language name is easier to identify for a person if it is directly displayed on their own language.