Added ideas to code for interaction system
NOTE : This system is probably going to become it's own component, maybe something like an InteractTriggerController or whatever. That is, if it gets implemented in the end, because at the current level of complexity, the game does not really need this to be a thing, which is why this code is commented out as of now.