Modified ConsoleUIController to have a console enabled property
The Console UI Controller now has a variable that allows the console to be set to enabled and disabled. This way, users wont be able to use the console accidentally unless they manually enable it in the settings menu.

Also, an issue where the isOpen parameter for the set console open function was never used. Whoops, funny oversight :P