Updated AccountManager to perform settings fetching in 2 steps
Yes, this is suboptimal and requires some more code changes, but it makes the code generic enough that it becomes extremely easy to implement settings loading from local files for native builds in the future.

Yes, this could be done better, but with deadlines I don't have time. Please forgive my crimes...