Reworked SettingsManager saving and loading system
The settings manager class now acts as a DTO handler manager class for settings. It generates an UserSettings class object when needed, collecting the information and settings from the real physical systems of the program and then sends them as JSON to the web server to be stored, acitng as a simple packaging DTO class.

It also has a method to set data from loaded data, which allows the system to modify the values of the real physical systems based on those of an UserSettings object given to it.

These changes make serialization and deserialization code generic enough that it should be easier to modify the systmen in the long run, as well as allowing an easier implementation for local file data management.