Modified FadeUIController to change opacity at constant rate rather than lerping
NOTE : The speed is constant from the point of view that the speed value is the same for a given duration, but the speed itself is not constant per frame since the framerate changes. That is why the value is multiplied by deltaTime, to ensure that it is consistent across different framerates.