Updated fade controller to use unscaled delta time
With this change, transitions should last the same no matter what the time scale is.

This may become an unwanted behaviour in the future if gameplay related events start to use time scale, but as of now it is a good change, since it allows the game to load scenes with transitions even when paused.