Modified TransitionToSceneInternal coroutine to use WaitForSecondsRealtime
The TransitionToSceneInternal coroutine handles loading scenes with a small delay for screen transitions.

This coroutine used to use the WaitForSeconds() class for the transition delay, which actually waits using scaled delta time. This meant that trying to load scenes while paused would not work at all until the game was unpaused, since pausing sets the time scale to 0.

This issue has been fixed by modifying the method to make use of the WaitForSecondsRealtime() class instead.