Modified code to only call Resume() if the game is paused
This is not really a fix, as the previous implementation did not have any issues. In any case, it feels cleaner to not just blindly call the unpause code every time we load a new scene, specially since Resume() could be modified in the future to add more code that we may not want to accidentally call every single time we load into a new scene.