Modified beam controller to have a child beam spawning flag
This flag controls if the beam can spawn child beams or not. This makes it easier to debug and spawn beams that have a different logic according to the situation.

Enabling "canMix" makes it possible for beams to generate child beams.

Disabling "canMix" makes it impossible for beams to generate child beams.