Added CanActivateTriggers property to player controller
Added a property to the player controller which determines whether the player can activate or not triggers.

This property will be used mainly for debugging purposes, but in the future, when it is made into a generic EntityController property, it will be possible to implement more interesting scenarios where all NPCs except one specific NPC can activate certain types of triggers.