Replaced texture with lower resolution version
Accidentally imported the internal 1024x1024 image rather than the 256x256 one. Remember, this is a web game, so better to be careful with this sort of stuff!

NOTE : In the future it could be very useful to reduce the number of materials and use atlases / textures that bundle multiple textures inside so that we can better use the available resources...