Created NPCAnimationController script
This script will control the logic for playing animations for NPCs

TODO : In the future, merge the logic with the animation controller of the player. That needs a refactor of the code to implement the whole entity system I had planned. And with deadlines, there is no time for that rewrite, so just deal with the crap for now.... agile my ass....